Move pending Blosc2 functions to service-specific utilities
