+++ caterva2/services/srv_utils.py
@@ -22,6 +22,38 @@ from caterva2 import models
 from caterva2 import api_utils
 
 
+#
+# Blosc2 related functions
+#
+
+def compress(data, dst=None):
+    assert isinstance(data, (bytes, pathlib.Path))
+
+    if dst is not None:
+        dst.parent.mkdir(exist_ok=True, parents=True)
+        if dst.exists():
+            dst.unlink()
+
+    # Create schunk
+    cparams = {}
+    dparams = {}
+    storage = {
+        'urlpath': dst,
+        'cparams': cparams,
+        'dparams': dparams,
+    }
+    schunk = blosc2.SChunk(**storage)
+
+    # Append data
+    if isinstance(data, pathlib.Path):
+        with open(data, 'rb') as f:
+            data = f.read()
+
+    schunk.append_data(data)
+
+    return schunk
+
+
 def open_b2(abspath):
     suffix = abspath.suffix
     if suffix == '.b2nd':
@@ -37,6 +69,19 @@ def open_b2(abspath):
         raise NotImplementedError()
 
     return array, schunk
+
+
+def chunk_is_available(schunk, nchunk):
+    # Blosc2 flags are at offset 31
+    # (see https://github.com/Blosc/c-blosc2/blob/main/README_CHUNK_FORMAT.rst)
+    flag = (schunk.get_lazychunk(nchunk)[31] & 0b01110000) >> 4
+    return flag != blosc2.SpecialValue.UNINIT.value
+
+
+def iterchunk(chunk):
+    # TODO Yield block by block
+    yield chunk
+
 
 def init_b2nd(metadata, urlpath=None):
     if urlpath is not None:
